Add files via upload
This is the modification of the original caldav2ics-plugin.php file from Werner Joss. This file provides the option entering the "exclude phrase" via the GUI at the admin page of the plugin. This file is NOT meant to replace the original file from W. Joss in his plugin WP-CALDAAV2ICS.